+++ ydb/tests/datashard/async_replication/test_async_replication.py
@@ -1,11 +1,12 @@
 import pytest
 import time
 from ydb.tests.datashard.lib.base_async_replication import TestBase
-from ydb.tests.datashard.lib.create_table import create_table, create_ttl, cleanup_type_name, pk_types, non_pk_types, index_first, index_second, ttl_types, \
+from ydb.tests.datashard.lib.dml import DML
+from ydb.tests.datashard.lib.types_of_variables import cleanup_type_name, format_sql_value, pk_types, non_pk_types, index_first, index_second, ttl_types, \
     index_first_sync, index_second_sync, index_three_sync, index_four_sync, index_zero_sync
 
 
-class TestAsyncReplication(TestBase):
+class TestAsyncReplication(TestBase, DML):
     @pytest.mark.parametrize(
         "table_name, pk_types, all_types, index, ttl, unique, sync",
         [
@@ -70,89 +71,32 @@ class TestAsyncReplication(TestBase):
             table_name, all_types, pk_types, index, ttl)
 
     def select_async_after_insert(self, table_name: str, all_types: dict[str, str], pk_types: dict[str, str], index: dict[str, str], ttl: str):
-        err = """"Type annotation" issue_code: 1030 severity: 1 issues { position { row: 2 column: 17 } message: "At function: KiReadTable!" end_position { row: 2 column: 17 }"""
-        err += """severity: 1 issues { position { row: 2 column: 17 } message:"""
-        err += f""" "Cannot find table \'db.[/Root/{table_name}]\' because it does not exist or you do not have access permissions."""
-        err += """ Please check correctness of table path and user permissions." end_position { row: 2 column: 17 } issue_code: 2003 severity: 1 } } (server_code: 400070)"""
+
         number_of_columns = len(pk_types) + len(all_types) + len(index)
 
         if ttl != "":
             number_of_columns += 1
 
         for count in range(1, number_of_columns + 1):
-            sql_select = f"""
-                SELECT COUNT(*) as count FROM `{table_name}` WHERE """
-
-            for type_name in pk_types.keys():
-                sql_select += f"pk_{cleanup_type_name(type_name)}={pk_types[type_name].format(count)} and "
+            create_all_type = []
             for type_name in all_types.keys():
                 if type_name != "Json" and type_name != "Yson" and type_name != "JsonDocument":
-                    sql_select += f"col_{cleanup_type_name(type_name)}={all_types[type_name].format(count)} and "
-            for type_name in index.keys():
-                sql_select += f"col_index_{cleanup_type_name(type_name)}={index[type_name].format(count)} and "
-            if ttl != "":
-                sql_select += f"ttl_{ttl}={ttl_types[ttl].format(count)}"
-            else:
-                sql_select += f"""pk_Int64={pk_types["Int64"].format(count)}"""
-            for _ in range(10):
-                try:
-                    rows = self.query_async(sql_select)
-                    break
-                except Exception:
-                    time.sleep(1)
-
-            for _ in range(10):
-                rows = self.query_async(sql_select)
-                if len(rows) == 1 and rows[0].count == 1:
-                    break
-                time.sleep(1)
+                    create_all_type.append(
+                        f"col_{cleanup_type_name(type_name)}={format_sql_value(all_types[type_name](count), type_name)}")
+            sql_select = f"""
+                SELECT COUNT(*) as count FROM `{table_name}` WHERE 
+                {" and ".join([f"pk_{cleanup_type_name(type_name)}={format_sql_value(pk_types[type_name](count), type_name)}" for type_name in pk_types.keys()])}
+                {" and " if len(index) != 0 else ""}
+                {" and ".join([f"col_index_{cleanup_type_name(type_name)}={format_sql_value(index[type_name](count), type_name)}" for type_name in index.keys()])}
+                {" and " if len(create_all_type) != 0 else ""}
+                {" and ".join(create_all_type)}
+                {f" and  ttl_{ttl}={format_sql_value(ttl_types[ttl](count), ttl)}" if ttl != "" else ""}
+                """
+            rows = self.query(sql_select)
             assert len(
                 rows) == 1 and rows[0].count == 1, f"Expected one rows, faild in {count} value, table {table_name}"
 
-    # После мерджа test_DML убрать эти методы https://github.com/ydb-platform/ydb/pull/16117/files
-    def create_table(self, table_name: str, pk_types: dict[str, str], all_types: dict[str, str], index: dict[str, str], ttl: str, unique: str, sync: str):
-        columns = {
-            "pk_": pk_types.keys(),
-            "col_": all_types.keys(),
-            "col_index_": index.keys(),
-            "ttl_": [ttl]
-        }
-        pk_columns = {
-            "pk_": pk_types.keys()
-        }
-        index_columns = {
-            "col_index_": index.keys()
-        }
-        sql_create_table = create_table(
-            table_name, columns, pk_columns, index_columns, unique, sync)
-        self.query(sql_create_table)
-        if ttl != "":
-            sql_ttl = create_ttl(f"ttl_{cleanup_type_name(ttl)}", {"P18262D": ""}, "SECONDS" if ttl ==
-                                 "Uint32" or ttl == "Uint64" or ttl == "DyNumber" else "", table_name)
-            self.query(sql_ttl)
-
-    def insert(self, table_name: str, all_types: dict[str, str], pk_types: dict[str, str], index: dict[str, str], ttl: str):
-        number_of_columns = len(pk_types) + len(all_types) + len(index)
-
-        if ttl != "":
-            number_of_columns += 1
-        for count in range(1, number_of_columns + 1):
-            self.create_insert(table_name, count, all_types,
-                               pk_types, index, ttl)
-
-    def create_insert(self, table_name: str, value: int, all_types: dict[str, str], pk_types: dict[str, str], index: dict[str, str], ttl: str):
-        insert_sql = f"""
-            INSERT INTO {table_name}(
-                {", ".join(["pk_" + cleanup_type_name(type_name) for type_name in pk_types.keys()])}{", " if len(all_types) != 0 else ""}
-                {", ".join(["col_" + cleanup_type_name(type_name) for type_name in all_types.keys()])}{", " if len(index) != 0 else ""}
-                {", ".join(["col_index_" + cleanup_type_name(type_name) for type_name in index.keys()])}{", " if len(ttl) != 0 else ""}
-                {f"ttl_{ttl}" if ttl != "" else ""}
-            )
-            VALUES(
-                {", ".join([pk_types[type_name].format(value) for type_name in pk_types.keys()])}{", " if len(all_types) != 0 else ""}
-                {", ".join([all_types[type_name].format(value) for type_name in all_types.keys()])}{", " if len(index) != 0 else ""}
-                {", ".join([index[type_name].format(value) for type_name in index.keys()])}{", " if len(ttl) != 0 else ""}
-                {ttl_types[ttl].format(value) if ttl != "" else ""}
-            );
-        """
-        self.query(insert_sql)
+        rows = self.query_async(
+            f"SELECT COUNT(*) as count FROM `{table_name}`")
+        assert len(
+            rows) == 1 and rows[0].count == number_of_columns, f"Expected {number_of_columns} rows, after select all line"
